+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- [FEATURE] Add support for view loading API (addViewLoadingTime). See [#2026][]
 - [IMPROVEMENT] Drop support for deprecated cocoapod specs. See [#1998][]
 - [FIX] Propagate global Tracer tags to OpenTelemetry span attributes. See [#2000][]
 - [FEATURE] Add Logs event mapper to ObjC API. See [#2008][]
@@ -758,6 +759,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#2005]: https://github.com/DataDog/dd-sdk-ios/pull/2005
 [#1998]: https://github.com/DataDog/dd-sdk-ios/pull/1998
 [#1998]: https://github.com/DataDog/dd-sdk-ios/pull/1966
+[#2026]: https://github.com/DataDog/dd-sdk-ios/pull/2026
 [@00fa9a]: https://github.com/00FA9A
 [@britton-earnin]: https://github.com/Britton-Earnin
 [@hengyu]: https://github.com/Hengyu